requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
auto-py-to-exe==2.48.1
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2026.1.4
charset-normalizer==3.4.4
distlib==0.4.0
distro==1.9.0
Eel==0.18.2
elevenlabs==2.34.0
filelock==3.20.3
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
jiter==0.13.0
openai==2.17.0
packaging==26.0
platformdirs==4.5.1
playsound3==3.3.1
playwright==1.58.0
proxy_tools==0.1.0
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dotenv==1.2.1
pywebview==6.1
QtPy==2.4.3
requests==2.32.5
setuptools==81.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
websockets==16.0
zope.event==6.1
zope.interface==8.2
